Fix 6 contraption assembly and movement bugs from NeoForge 6.0.9 audit
1. Contraption.moveSeat(): Use AllBlockTags.SEATS tag instead of instanceof
   SeatBlock - modded/addon seats now correctly captured by contraptions
2. Contraption.moveBlock(): Add ChainConveyorBlockEntity.notifyConnectedToValidate()
   call when chain conveyors are assembled into contraptions
3. ContraptionCollider: Add tickRateManager().isEntityFrozen() check to skip
   collision with frozen entities (command-frozen or tick rate manager)
4. ContraptionCollider: Contract client player bounds by 2/16 blocks when taller
   than 1 block to prevent getting stuck in contraptions
5. ControlledContraptionEntity.applyLocalTransforms(): Add null check on rotation
   axis to prevent crash for translating contraptions (e.g. elevators)
6. MountedFluidStorage.deserialize(): Read capacity as long instead of int to
   match serialization format and prevent truncation of large tank capacities

+++ modules/create/src/main/java/com/simibubi/create/content/contraptions/MountedFluidStorage.java
@@ -146,7 +146,7 @@ public class MountedFluidStorage {
 		if (nbt == null)
 			return storage;
 
-		int capacity = nbt.getInt("Capacity");
+		long capacity = nbt.getLong("Capacity");
 		storage.tank = new SmartFluidTank(capacity, storage::onFluidStackChanged);
 		storage.valid = true;
